Add Wallaby
Looks better than hound. People are already using it for one of the
things we want to use it for (web socket testing). DSL seems friendlier
and more expressive. Browser state is passed around explicitly in
session, whereas Hound seems to be all side effects. We’re not totally
free from them here (session can be mutated without a return value) but
overall it seems more in keeping with the spirit of the ecosystem we’re
working in

We can always go back